Fix SonarCloud gate issues
