+++ packages/sync/src/source/source.ts
@@ -23,6 +23,7 @@ import {
   resolveEntityFunctionKey,
   resolveDefaultUnitValue,
   resolveInput,
+  resolveSingleInFlight,
   resolveUnitRunAsVoid,
   resolveUnitMode,
   resolveValue,
@@ -750,23 +751,6 @@ export const createSourceFromConfig = <
     let singleInFlightPayloadKey: string | null = null;
     let runConfig: SourceRunConfig | undefined;
 
-    const resolveSingleInFlightPromise = (): Promise<TData> | null => {
-      if (!singleInFlightPromise) {
-        return null;
-      }
-
-      if (hasSingleInFlightPayload && Object.is(singleInFlightPayload, internal.payload)) {
-        return singleInFlightPromise;
-      }
-
-      const payloadKey = payloadKeyCache.getOrCreateKey(internal.payload);
-      if (singleInFlightPayloadKey === null && hasSingleInFlightPayload) {
-        singleInFlightPayloadKey = payloadKeyCache.getOrCreateKey(singleInFlightPayload);
-      }
-
-      return singleInFlightPayloadKey === payloadKey ? singleInFlightPromise : null;
-    };
-
     const clearInFlightTracking = (): void => {
       internal.inFlightByPayload.clear();
       singleInFlightPromise = null;
@@ -825,9 +809,18 @@ export const createSourceFromConfig = <
         internal.payload = nextPayload;
       }
       let payloadKey: string | null = null;
-      const matchingSingleInFlightPromise = resolveSingleInFlightPromise();
-      if (matchingSingleInFlightPromise) {
-        return matchingSingleInFlightPromise;
+      const singleInFlight = resolveSingleInFlight({
+        currentPayload: internal.payload,
+        hasTrackedPayload: hasSingleInFlightPayload,
+        payloadKeyCache,
+        promise: singleInFlightPromise,
+        trackedPayload: singleInFlightPayload,
+        trackedPayloadKey: singleInFlightPayloadKey,
+      });
+      singleInFlightPayloadKey = singleInFlight.trackedPayloadKey;
+      payloadKey = singleInFlight.currentPayloadKey;
+      if (singleInFlight.promise) {
+        return singleInFlight.promise;
       }
 
       if (internal.inFlightByPayload.size > 0) {
