+++ packages/sync/src/sourceLazy.ts
@@ -177,9 +177,14 @@ const createLazySourceFromConfig = <
     ) => {
       const unit = await ensureUnit();
       const load = unit.getSnapshot().load;
-      const args = configOrMode === undefined
-        ? (dataOrSetAction === undefined ? [] : [dataOrSetAction])
-        : [dataOrSetAction, configOrMode];
+      const args: unknown[] = [];
+      if (configOrMode === undefined) {
+        if (dataOrSetAction !== undefined) {
+          args.push(dataOrSetAction);
+        }
+      } else {
+        args.push(dataOrSetAction, configOrMode);
+      }
 
       await Reflect.apply(load, unit, args);
