+++ sonar-project.properties
@@ -2,7 +2,10 @@ sonar.organization=live-input-vector-output-node
 sonar.projectKey=live-input-vector-output-node_livon-ts
 sonar.projectName=livon-ts
 sonar.sources=.
+sonar.tests=.
+sonar.test.inclusions=**/*.spec.ts,**/*.spec.tsx,**/*.test.ts,**/*.test.tsx
 sonar.qualitygate.wait=true
 sonar.qualitygate.timeout=600
 sonar.exclusions=**/node_modules/**,**/.turbo/**,**/dist/**,**/build/**,**/coverage/**,**/.next/**,**/.pnpm-store/**,**/.scannerwork/**,**/*.png,**/*.jpg,**/*.jpeg,**/*.ico
 sonar.coverage.exclusions=.github/**,tools/ci/**,tools/policies/**,**/*.config.*,**/*.d.ts
+sonar.javascript.lcov.reportPaths=coverage/lcov.info
